HTTPClient: add `async` overload for `execute` function
Additionally, moved `HTTPClientHeaders`, `HTTPClientRequest`, and `HTTPClientResponse` to separate files for easier maintenance.
