+++ Sources/Basics/HTTPClient/HTTPClient.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift open source project
 //
-// Copyright (c) 2020 Apple Inc. and the Swift project authors
+// Copyright (c) 2020-2023 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See http://swift.org/LICENSE.txt for license information
@@ -344,217 +344,4 @@ public enum HTTPClientRetryStrategy {
 
 public enum HTTPClientCircuitBreakerStrategy {
     case hostErrors(maxErrors: Int, age: DispatchTimeInterval)
-}
-
-// MARK: - HTTPClientRequest
-
-public struct HTTPClientRequest {
-    public let kind: Kind
-    public let url: URL
-    public var headers: HTTPClientHeaders
-    public var body: Data?
-    public var options: Options
-
-    public init(kind: Kind,
-                url: URL,
-                headers: HTTPClientHeaders = .init(),
-                body: Data? = nil,
-                options: Options = .init()) {
-        self.kind = kind
-        self.url = url
-        self.headers = headers
-        self.body = body
-        self.options = options
-    }
-
-    // generic request
-    public init(method: Method = .get,
-                url: URL,
-                headers: HTTPClientHeaders = .init(),
-                body: Data? = nil,
-                options: Options = .init()) {
-        self.init(kind: .generic(method), url: url, headers: headers, body: body, options: options)
-    }
-
-    // download request
-    public static func download(url: URL,
-                                headers: HTTPClientHeaders = .init(),
-                                options: Options = .init(),
-                                fileSystem: FileSystem,
-                                destination: AbsolutePath) -> HTTPClientRequest {
-        HTTPClientRequest(kind: .download(fileSystem: fileSystem, destination: destination),
-                          url: url,
-                          headers: headers,
-                          body: nil,
-                          options: options)
-    }
-
-    public var method: Method {
-        switch self.kind {
-        case .generic(let method):
-            return method
-        case .download:
-            return .get
-        }
-    }
-
-    public enum Kind {
-        case generic(Method)
-        case download(fileSystem: FileSystem, destination: AbsolutePath)
-    }
-
-    public enum Method {
-        case head
-        case get
-        case post
-        case put
-        case delete
-    }
-
-    public struct Options {
-        public var addUserAgent: Bool
-        public var validResponseCodes: [Int]?
-        public var timeout: DispatchTimeInterval?
-        public var maximumResponseSizeInBytes: Int64?
-        public var authorizationProvider: HTTPClientAuthorizationProvider?
-        public var retryStrategy: HTTPClientRetryStrategy?
-        public var circuitBreakerStrategy: HTTPClientCircuitBreakerStrategy?
-        public var callbackQueue: DispatchQueue?
-
-        public init() {
-            self.addUserAgent = true
-            self.validResponseCodes = .none
-            self.timeout = .none
-            self.maximumResponseSizeInBytes = .none
-            self.authorizationProvider = .none
-            self.retryStrategy = .none
-            self.circuitBreakerStrategy = .none
-            self.callbackQueue = .none
-        }
-    }
-}
-
-// MARK: - HTTPClientResponse
-
-public struct HTTPClientResponse {
-    public let statusCode: Int
-    public let statusText: String?
-    public let headers: HTTPClientHeaders
-    public let body: Data?
-
-    public init(statusCode: Int,
-                statusText: String? = nil,
-                headers: HTTPClientHeaders = .init(),
-                body: Data? = nil) {
-        self.statusCode = statusCode
-        self.statusText = statusText
-        self.headers = headers
-        self.body = body
-    }
-
-    public func decodeBody<T: Decodable>(_ type: T.Type, using decoder: JSONDecoder = .init()) throws -> T? {
-        try self.body.flatMap { try decoder.decode(type, from: $0) }
-    }
-}
-
-extension HTTPClientResponse {
-    public static func okay(body: String? = nil) -> HTTPClientResponse {
-        return .okay(body: body?.data(using: .utf8))
-    }
-
-    public static func okay(body: Data?) -> HTTPClientResponse {
-        return HTTPClientResponse(statusCode: 200, body: body)
-    }
-
-    public static func notFound(reason: String? = nil) -> HTTPClientResponse {
-        return HTTPClientResponse(statusCode: 404, body: (reason ?? "Not Found").data(using: .utf8))
-    }
-
-    public static func serverError(reason: String? = nil) -> HTTPClientResponse {
-        return HTTPClientResponse(statusCode: 500, body: (reason ?? "Internal Server Error").data(using: .utf8))
-    }
-}
-
-// MARK: - HTTPClientHeaders
-
-public struct HTTPClientHeaders {
-    private var items: [Item]
-    private var headers: [String: [String]]
-
-    public init(_ items: [Item] = []) {
-        self.items = items
-        self.headers = items.reduce([String: [String]]()) { partial, item in
-            var map = partial
-            // Avoid copy-on-write: remove entry from dictionary before mutating
-            var values = map.removeValue(forKey: item.name.lowercased()) ?? []
-            values.append(item.value)
-            map[item.name.lowercased()] = values
-            return map
-        }
-    }
-
-    public func contains(_ name: String) -> Bool {
-        self.headers[name.lowercased()] != nil
-    }
-
-    public var count: Int {
-        self.headers.count
-    }
-
-    public mutating func add(name: String, value: String) {
-        self.add(Item(name: name, value: value))
-    }
-
-    public mutating func add(_ item: Item) {
-        self.add([item])
-    }
-
-    public mutating func add(_ items: [Item]) {
-        for item in items {
-            if self.items.contains(item) {
-                continue
-            }
-            // Avoid copy-on-write: remove entry from dictionary before mutating
-            var values = self.headers.removeValue(forKey: item.name.lowercased()) ?? []
-            values.append(item.value)
-            self.headers[item.name.lowercased()] = values
-            self.items.append(item)
-        }
-    }
-
-    public mutating func merge(_ other: HTTPClientHeaders) {
-        self.add(other.items)
-    }
-
-    public func get(_ name: String) -> [String] {
-        self.headers[name.lowercased()] ?? []
-    }
-
-    public struct Item: Equatable {
-        let name: String
-        let value: String
-
-        public init(name: String, value: String) {
-            self.name = name
-            self.value = value
-        }
-    }
-}
-
-extension HTTPClientHeaders: Sequence {
-    public func makeIterator() -> IndexingIterator<[Item]> {
-        return self.items.makeIterator()
-    }
-}
-
-extension HTTPClientHeaders: Equatable {
-    public static func == (lhs: HTTPClientHeaders, rhs: HTTPClientHeaders) -> Bool {
-        return lhs.headers == rhs.headers
-    }
-}
-
-extension HTTPClientHeaders: ExpressibleByDictionaryLiteral {
-    public init(dictionaryLiteral elements: (String, String)...) {
-        self.init(elements.map(Item.init))
-    }
 }
