+++ Sources/Basics/HTTPClient/HTTPClientRequest.swift
@@ -1,0 +1,102 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift open source project
+//
+// Copyright (c) 2020-2023 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+import Foundation
+
+import protocol TSCBasic.FileSystem
+import struct TSCBasic.AbsolutePath
+
+public struct HTTPClientRequest {
+    public let kind: Kind
+    public let url: URL
+    public var headers: HTTPClientHeaders
+    public var body: Data?
+    public var options: Options
+
+    public init(kind: Kind,
+                url: URL,
+                headers: HTTPClientHeaders = .init(),
+                body: Data? = nil,
+                options: Options = .init()) {
+        self.kind = kind
+        self.url = url
+        self.headers = headers
+        self.body = body
+        self.options = options
+    }
+
+    // generic request
+    public init(method: Method = .get,
+                url: URL,
+                headers: HTTPClientHeaders = .init(),
+                body: Data? = nil,
+                options: Options = .init()) {
+        self.init(kind: .generic(method), url: url, headers: headers, body: body, options: options)
+    }
+
+    // download request
+    public static func download(url: URL,
+                                headers: HTTPClientHeaders = .init(),
+                                options: Options = .init(),
+                                fileSystem: FileSystem,
+                                destination: AbsolutePath) -> HTTPClientRequest {
+        HTTPClientRequest(kind: .download(fileSystem: fileSystem, destination: destination),
+                          url: url,
+                          headers: headers,
+                          body: nil,
+                          options: options)
+    }
+
+    public var method: Method {
+        switch self.kind {
+        case .generic(let method):
+            return method
+        case .download:
+            return .get
+        }
+    }
+
+    public enum Kind {
+        case generic(Method)
+        case download(fileSystem: FileSystem, destination: AbsolutePath)
+    }
+
+    public enum Method {
+        case head
+        case get
+        case post
+        case put
+        case delete
+    }
+
+    public struct Options {
+        public var addUserAgent: Bool
+        public var validResponseCodes: [Int]?
+        public var timeout: DispatchTimeInterval?
+        public var maximumResponseSizeInBytes: Int64?
+        public var authorizationProvider: HTTPClientAuthorizationProvider?
+        public var retryStrategy: HTTPClientRetryStrategy?
+        public var circuitBreakerStrategy: HTTPClientCircuitBreakerStrategy?
+        public var callbackQueue: DispatchQueue?
+
+        public init() {
+            self.addUserAgent = true
+            self.validResponseCodes = .none
+            self.timeout = .none
+            self.maximumResponseSizeInBytes = .none
+            self.authorizationProvider = .none
+            self.retryStrategy = .none
+            self.circuitBreakerStrategy = .none
+            self.callbackQueue = .none
+        }
+    }
+}
